fix: remove "size" attribute from view()
In most cases this size attribute is much too small and results in the
FST being cut off at the right and bottom, often drastically so, when
viewed in Jupyter.  If we simply do not set "size" then it might not
fit in the cell but at least we get scroll bars.  Also the user can
set "size" after the fact, e.g.:

    dg = fst.view()
    dg.attr(size, "42,14")
    dg

but it is less obvious how to *remove* the size (I think it can be
done with `del dg.graph_attr["size"]`)